feat: #21. 뚜벗 뽑기 후 이름 등록 api 설꼐

+++ server/userttubeot/src/main/java/com/user/userttubeot/ttubeot/domain/dto/TtubeotNameRegisterRequestDTO.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class TtubeotNameRegisterRequestDTO {
 
-    private Integer userTtubeotOwnershipId;
+    private Long userTtubeotOwnershipId;
     private String userTtubeotOwnershipName; // 등록할 뚜벗의 이름
 
 }
